Refactored code to support different types of kubernetes objects creation through Twill interface.
