+++ cdap-kubernetes/src/main/java/io/cdap/cdap/k8s/runtime/AbstractKubeTwillRunnerService.java
@@ -17,12 +17,10 @@
 package io.cdap.cdap.k8s.runtime;
 
 import com.google.common.util.concurrent.Uninterruptibles;
+import io.cdap.cdap.k8s.common.KubeResourceType;
 import io.cdap.cdap.k8s.common.ResourceChangeListener;
 import io.cdap.cdap.master.environment.k8s.PodInfo;
 import io.kubernetes.client.ApiClient;
-import io.kubernetes.client.models.V1Deployment;
-import io.kubernetes.client.models.V1DeploymentCondition;
-import io.kubernetes.client.models.V1DeploymentStatus;
 import io.kubernetes.client.models.V1ObjectMeta;
 import io.kubernetes.client.util.Config;
 import org.apache.twill.api.ResourceSpecification;
@@ -46,7 +44,6 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
@@ -77,19 +74,15 @@ import javax.annotation.Nullable;
  *
  * cdap.twill.app=[literal app name]
  */
-public class KubeTwillRunnerService implements TwillRunnerService {
+abstract class AbstractKubeTwillRunnerService<T> implements TwillRunnerService {
+  private static final Logger LOG = LoggerFactory.getLogger(AbstractKubeTwillRunnerService.class);
 
   static final String START_TIMEOUT_ANNOTATION = "cdap.app.start.timeout.millis";
-
-  private static final Logger LOG = LoggerFactory.getLogger(KubeTwillRunnerService.class);
-
   private static final String APP_ANNOTATION = "cdap.twill.app";
   private static final String APP_LABEL = "cdap.twill.app";
   private static final String RUNNER_LABEL = "cdap.twill.runner";
   private static final String RUNNER_LABEL_VAL = "k8s";
   private static final String RUN_ID_LABEL = "cdap.twill.run.id";
-
-  private static final String AVAILABLE_TYPE = "Available";
 
   private final String kubeNamespace;
   private final String resourcePrefix;
@@ -97,15 +90,16 @@ public class KubeTwillRunnerService implements TwillRunnerService {
   private final PodInfo podInfo;
   private final DiscoveryServiceClient discoveryServiceClient;
   private final Map<String, String> extraLabels;
-  private final DeploymentWatcher deploymentWatcher;
+  private final KubeResourceWatcher<T> kubeResourceWatcher;
   private final Map<String, KubeLiveInfo> liveInfos;
   private final Lock liveInfoLock;
   private ScheduledExecutorService monitorScheduler;
   private ApiClient apiClient;
 
-  public KubeTwillRunnerService(String kubeNamespace, DiscoveryServiceClient discoveryServiceClient,
-                                PodInfo podInfo, String resourcePrefix, Map<String, String> cConf,
-                                Map<String, String> extraLabels) {
+  AbstractKubeTwillRunnerService(String kubeNamespace, DiscoveryServiceClient discoveryServiceClient,
+                                 PodInfo podInfo, String resourcePrefix, Map<String, String> cConf,
+                                 Map<String, String> extraLabels, KubeResourceType kubeResourceType) {
+
     this.kubeNamespace = kubeNamespace;
     this.podInfo = podInfo;
     this.resourcePrefix = resourcePrefix;
@@ -113,9 +107,8 @@ public class KubeTwillRunnerService implements TwillRunnerService {
     this.discoveryServiceClient = discoveryServiceClient;
     this.extraLabels = Collections.unmodifiableMap(new HashMap<>(extraLabels));
     // Selects all runs start by the k8s twill runner that has the run id label
-    this.deploymentWatcher = new DeploymentWatcher(kubeNamespace, String.format("%s=%s,%s",
-                                                                                RUNNER_LABEL, RUNNER_LABEL_VAL,
-                                                                                RUN_ID_LABEL));
+    String selector = String.format("%s=%s,%s", RUNNER_LABEL, RUNNER_LABEL_VAL, RUN_ID_LABEL);
+    this.kubeResourceWatcher = new KubeResourceWatcher<>(kubeNamespace, selector, kubeResourceType);
     this.liveInfos = new ConcurrentSkipListMap<>();
     this.liveInfoLock = new ReentrantLock();
   }
@@ -145,8 +138,9 @@ public class KubeTwillRunnerService implements TwillRunnerService {
     // labels have more strict requirements around valid character sets,
     // so use annotations to store the app name.
     resourceMeta.setAnnotations(Collections.singletonMap(APP_ANNOTATION, spec.getName()));
-    return new KubeTwillPreparer(apiClient, kubeNamespace, podInfo,
-                                 spec, runId, resourceMeta, cConf, (timeout, timeoutUnit) -> {
+    return getTwillPreparer(apiClient, kubeNamespace, podInfo,
+                            spec, runId, resourceMeta, cConf, (timeout, timeoutUnit) -> {
+
       // Adds the controller to the LiveInfo.
       liveInfoLock.lock();
       try {
@@ -159,6 +153,14 @@ public class KubeTwillRunnerService implements TwillRunnerService {
       }
     });
   }
+
+  abstract TwillPreparer getTwillPreparer(ApiClient apiClient, String kubeNamespace, PodInfo podInfo,
+                                          TwillSpecification spec, RunId twillRunId, V1ObjectMeta resourceMeta,
+                                          Map<String, String> cConf, KubeTwillControllerFactory controllerFactory);
+
+  abstract V1ObjectMeta getObjectMeta(T resource);
+
+  abstract boolean isObjectAvailable(T resource);
 
   @Nullable
   @Override
@@ -197,8 +199,8 @@ public class KubeTwillRunnerService implements TwillRunnerService {
       apiClient = Config.defaultClient();
       monitorScheduler = Executors.newSingleThreadScheduledExecutor(
         Threads.createDaemonThreadFactory("kube-monitor-executor"));
-      deploymentWatcher.addListener(new DeploymentListener());
-      deploymentWatcher.start();
+      kubeResourceWatcher.addListener(new KubeObjectListener());
+      kubeResourceWatcher.start();
     } catch (IOException e) {
       throw new IllegalStateException("Unable to get Kubernetes API Client", e);
     }
@@ -206,7 +208,7 @@ public class KubeTwillRunnerService implements TwillRunnerService {
 
   @Override
   public void stop() {
-    deploymentWatcher.close();
+    kubeResourceWatcher.close();
     monitorScheduler.shutdownNow();
   }
 
@@ -235,32 +237,17 @@ public class KubeTwillRunnerService implements TwillRunnerService {
     // Listen to deployment changes. If the deployment represented by the controller changed to available, cancel
     // the terminationFuture. If the deployment is deleted, also cancel the terminationFuture, and also terminate
     // the controller.
-    Cancellable cancellable = deploymentWatcher.addListener(new ResourceChangeListener<V1Deployment>() {
+    Cancellable cancellable = kubeResourceWatcher.addListener(new ResourceChangeListener<T>() {
       @Override
-      public void resourceAdded(V1Deployment deployment) {
+      public void resourceAdded(T resource) {
         // Handle the same way as modified
-        resourceModified(deployment);
+        resourceModified(resource);
       }
 
       @Override
-      public void resourceModified(V1Deployment deployment) {
-        if (runId.equals(deployment.getMetadata().getLabels().get(RUN_ID_LABEL))) {
-          V1DeploymentStatus status = deployment.getStatus();
-          if (status == null) {
-            return;
-          }
-          List<V1DeploymentCondition> conditions = status.getConditions();
-          if (conditions == null) {
-            return;
-          }
-          boolean available = conditions.stream()
-            .filter(c -> AVAILABLE_TYPE.equals(c.getType()))
-            .map(V1DeploymentCondition::getStatus)
-            .findFirst()
-            .map(Boolean::parseBoolean)
-            .orElse(false);
-
-          if (available) {
+      public void resourceModified(T resource) {
+        if (runId.equals(getObjectMeta(resource).getLabels().get(RUN_ID_LABEL))) {
+          if (isObjectAvailable(resource)) {
             LOG.debug("Deployment for application {} with run {} is available", liveInfo.getApplicationName(), runId);
             // Cancel the scheduled termination
             terminationFuture.cancel(false);
@@ -275,9 +262,9 @@ public class KubeTwillRunnerService implements TwillRunnerService {
       }
 
       @Override
-      public void resourceDeleted(V1Deployment deployment) {
+      public void resourceDeleted(T resource) {
         // If the run is deleted, terminate the controller right away and cancel the scheduled termination
-        if (runId.equals(deployment.getMetadata().getLabels().get(RUN_ID_LABEL))) {
+        if (runId.equals(getObjectMeta(resource).getLabels().get(RUN_ID_LABEL))) {
           // Cancel the scheduled termination
           terminationFuture.cancel(false);
           controller.terminate();
@@ -318,22 +305,23 @@ public class KubeTwillRunnerService implements TwillRunnerService {
    * A {@link ResourceChangeListener} to watch for changes in all deployments. It is for refreshing the
    * liveInfos map.
    */
-  private final class DeploymentListener implements ResourceChangeListener<V1Deployment> {
+  private final class KubeObjectListener implements ResourceChangeListener<T> {
 
     @Override
-    public void resourceAdded(V1Deployment deployment) {
-      String appName = deployment.getMetadata().getAnnotations().get(APP_ANNOTATION);
+    public void resourceAdded(T resource) {
+      V1ObjectMeta objectMeta = getObjectMeta(resource);
+      String appName = objectMeta.getAnnotations().get(APP_ANNOTATION);
       if (appName == null) {
         // This shouldn't happen. Just to guard against future bug.
         return;
       }
-      String deploymentName = deployment.getMetadata().getName();
-      RunId runId = RunIds.fromString(deployment.getMetadata().getLabels().get(RUN_ID_LABEL));
+      String deploymentName = objectMeta.getName();
+      RunId runId = RunIds.fromString(objectMeta.getLabels().get(RUN_ID_LABEL));
 
-      // Read the start timeout millis from the annotation. It is set by the KubeTwillPreparer
+      // Read the start timeout millis from the annotation. It is set by the DeploymentTwillPreparer
       long startTimeoutMillis = TimeUnit.SECONDS.toMillis(120);
       try {
-        startTimeoutMillis = Long.parseLong(deployment.getMetadata().getAnnotations().get(START_TIMEOUT_ANNOTATION));
+        startTimeoutMillis = Long.parseLong(objectMeta.getAnnotations().get(START_TIMEOUT_ANNOTATION));
       } catch (Exception e) {
         // This shouldn't happen
         LOG.warn("Failed to get start timeout from the deployment annotation using key {}. Defaulting to {} ms",
@@ -353,8 +341,9 @@ public class KubeTwillRunnerService implements TwillRunnerService {
     }
 
     @Override
-    public void resourceDeleted(V1Deployment deployment) {
-      String appName = deployment.getMetadata().getAnnotations().get(APP_ANNOTATION);
+    public void resourceDeleted(T resource) {
+      V1ObjectMeta objectMeta = getObjectMeta(resource);
+      String appName = objectMeta.getAnnotations().get(APP_ANNOTATION);
       if (appName == null) {
         // This shouldn't happen. Just to guard against future bug.
         return;
@@ -365,7 +354,7 @@ public class KubeTwillRunnerService implements TwillRunnerService {
       try {
         KubeLiveInfo liveInfo = liveInfos.get(appName);
         if (liveInfo != null) {
-          RunId runId = RunIds.fromString(deployment.getMetadata().getLabels().get(RUN_ID_LABEL));
+          RunId runId = RunIds.fromString(objectMeta.getLabels().get(RUN_ID_LABEL));
           Optional.ofNullable(liveInfo.getController(runId)).ifPresent(KubeTwillController::terminate);
         }
       } finally {
