+++ cdap-kubernetes/src/main/java/io/cdap/cdap/master/environment/k8s/KubeMasterEnvironment.java
@@ -19,7 +19,7 @@ package io.cdap.cdap.master.environment.k8s;
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableSet;
 import io.cdap.cdap.k8s.discovery.KubeDiscoveryService;
-import io.cdap.cdap.k8s.runtime.KubeTwillRunnerService;
+import io.cdap.cdap.k8s.runtime.DeploymentTwillRunnerService;
 import io.cdap.cdap.master.spi.environment.MasterEnvironment;
 import io.cdap.cdap.master.spi.environment.MasterEnvironmentContext;
 import io.cdap.cdap.master.spi.environment.MasterEnvironmentTask;
@@ -93,7 +93,7 @@ public class KubeMasterEnvironment implements MasterEnvironment {
 
   private KubeDiscoveryService discoveryService;
   private PodKillerTask podKillerTask;
-  private KubeTwillRunnerService twillRunner;
+  private TwillRunnerService twillRunner;
 
   @Override
   public void initialize(MasterEnvironmentContext context) throws IOException, ApiException {
@@ -149,8 +149,8 @@ public class KubeMasterEnvironment implements MasterEnvironment {
                namespace, podKillerSelector, delayMillis);
     }
 
-    twillRunner = new KubeTwillRunnerService(namespace, discoveryService, podInfo, resourcePrefix, conf,
-                                             Collections.singletonMap(instanceLabel, instanceName));
+    twillRunner = new DeploymentTwillRunnerService(namespace, discoveryService, podInfo, resourcePrefix, conf,
+                                                   Collections.singletonMap(instanceLabel, instanceName));
     LOG.info("Kubernetes environment initialized with pod labels {}", podLabels);
   }
